Fix typo in test package name
